% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_longest_spell.R
\name{annual_longest_spell}
\alias{annual_longest_spell}
\title{Annual Longest Spell}
\usage{
annual_longest_spell(
  data,
  element,
  station = NULL,
  date = NULL,
  year = NULL,
  spell_type = c("between", "lte", "gte", "excluding between"),
  lower = 0,
  upper = 0.85,
  doy = NULL,
  doy_first = 1,
  doy_last = 366,
  result_name = "max_spell",
  na.rm = FALSE
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{element}{The name of the column in \code{data} to apply the spell condition to}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.
Spells are calculated separately for each station.}

\item{date}{The name of the date column in \code{data}. This is only needed if \code{year} is not specified as it will be created from \code{date}.}

\item{year}{The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{spell_type}{The spell condition type. One of "between", "lte" (less than or equal to),
"gte" (greater than or equal to) or "outside". See details for explanation.}

\item{lower}{The lower value of the spell condition (does not apply to spell type = "gte").}

\item{upper}{The lower value of the spell condition (does not apply to spell type = "lte").}

\item{doy}{The name of the day of the year (1-366) column in \code{data} if filtering to a part of the year.
Note that this filter is applied after calculating spell lengths, hence the longest spell could have started outside of this filter.
To exclude a part of the year from the calculations completely, filter the data before using \code{annual_longest_spell}.
If \code{doy} is \code{NULL} then it can be calculated as \code{yday_366(data[[date]])} if \code{date} is provided.}

\item{doy_first}{The first day of the year to consider the longest spell within.}

\item{doy_last}{The last day of the year to consider the longest spell within.}

\item{result_name}{The name of column of the length of the longest spell period in the resulting summary dataset.}

\item{na.rm}{A logical value. Should \code{NA} values be removed from the spell lengths before calculating the longest spell length?
Note that \code{NA} values in the element column are never removed to prevent spell runs continuing over non consecutive rows.}
}
\value{
A summary data.frame of the annual longest spell (for each station).
}
\description{
Calculates the length of the longest period each year for which the spell condition is valid
}
