% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_extremes.R
\name{climatic_extremes}
\alias{climatic_extremes}
\title{Calculate extremes}
\usage{
climatic_extremes(
  data,
  date_time,
  station = NULL,
  elements,
  year = NULL,
  month = NULL,
  dekad = NULL,
  pentad = NULL,
  to = c("hourly", "daily", "pentad", "dekadal", "monthly", "annual-within-year",
    "annual", "longterm-monthly", "longterm-within-year", "station", "overall"),
  by = NULL,
  doy = NULL,
  doy_first = 1,
  doy_last = 366,
  max_val = TRUE,
  min_val = FALSE,
  first_date = FALSE,
  n_dates = FALSE,
  last_date = FALSE,
  na.rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL,
  names = "{.fn}_{.col}"
)
}
\arguments{
\item{data}{}

\item{date_time}{}

\item{station}{}

\item{elements}{}

\item{year}{}

\item{month}{}

\item{dekad}{}

\item{pentad}{}

\item{to}{}

\item{by}{}

\item{doy}{}

\item{doy_first}{}

\item{doy_last}{}

\item{max_val}{If \code{TRUE} calculate the extreme maximum}

\item{min_val}{If \code{TRUE} calculate the extreme minimum}

\item{first_date}{If \code{TRUE} the first instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. first \code{date_time} when the maximum occurred.}

\item{n_dates}{If \code{TRUE} the number of \code{date_time} points when the value equals the summary value is included. Generally only used for extreme summaries i.e. number of days in which the minimum occurred.}

\item{last_date}{If \code{TRUE} the last instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. last \code{date_time} when the maximum occurred.}

\item{na.rm}{If \code{TRUE} all \code{na_} parameters are ignored and missing
values are removed. If \code{FALSE} missing values are not removed unless
any \code{na_} parameters are specified.}

\item{na_prop}{Max proportion of missing values allowed}

\item{na_n}{Max number of missing values allowed}

\item{na_consec}{Max number of consecutive missing values allowed}

\item{na_n_non}{Min number of non-missing values required}

\item{names}{Format of column names. Passed to \code{.names} in
\code{dplyr::across}}
}
\value{

}
\description{
Calculate extremes
}
