% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.R
\name{climdex}
\alias{climdex}
\title{Title}
\usage{
climdex(
  data,
  station,
  date,
  year,
  month,
  prec,
  tmax,
  tmin,
  indices,
  freq = "annual",
  base.range = c(1961, 1990),
  n = 5,
  northern.hemisphere = TRUE,
  quantiles = NULL,
  temp.qtiles = c(0.1, 0.9),
  prec.qtiles = c(0.95, 0.99),
  max.missing.days = c(annual = 15, monthly = 3),
  min.base.data.fraction.present = 0.1,
  spells.can.span.years = FALSE,
  gsl.mode = "GSL",
  threshold = 1
)
}
\arguments{
\item{data}{}

\item{station}{}

\item{date}{}

\item{year}{}

\item{month}{}

\item{prec}{}

\item{tmax}{}

\item{tmin}{}

\item{indices}{}

\item{freq}{}

\item{base.range}{}

\item{n}{}

\item{northern.hemisphere}{}

\item{quantiles}{}

\item{temp.qtiles}{}

\item{prec.qtiles}{}

\item{max.missing.days}{}

\item{min.base.data.fraction.present}{}

\item{spells.can.span.years}{}

\item{gsl.mode}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
