% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex_single_station.R
\name{climdex_single_station}
\alias{climdex_single_station}
\title{Title}
\usage{
climdex_single_station(
  ci,
  freq = "annual",
  indices,
  year,
  month,
  spells.can.span.years = FALSE,
  gsl.mode = gsl.mode,
  threshold = 1
)
}
\arguments{
\item{ci}{}

\item{freq}{}

\item{indices}{}

\item{year}{}

\item{month}{}

\item{spells.can.span.years}{}

\item{gsl.mode}{}

\item{threshold}{}
}
\value{

}
\description{
Title
}
