% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dec_deg.R
\name{convert_to_dec_deg}
\alias{convert_to_dec_deg}
\title{Title}
\usage{
convert_to_dec_deg(dd, mm = 0, ss = 0, dir = c("n", "e", "s", "w"))
}
\arguments{
\item{dd}{TODO}

\item{mm}{TODO}

\item{ss}{TODO}

\item{dir}{Direction to --- to. Takes values \code{"n"}, \code{"e"}, \code{"s"}, or \code{"w"}}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
