% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dekad.R
\name{dekad}
\alias{dekad}
\title{Get the dekad component of a date-time}
\usage{
dekad(date)
}
\arguments{
\item{date}{A date-time object}
}
\value{

}
\description{
Get the dekad component of a date-time
}
\examples{
dekad(as.Date("2020/12/25"))
dekad(as.Date("1999/01/01"))
}
