% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdt_dekad.R
\name{export_cdt_dekad}
\alias{export_cdt_dekad}
\title{Export data in CDT format for dekad data}
\usage{
export_cdt_dekad(
  data,
  station,
  element,
  date,
  latitude,
  longitude,
  altitude,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL,
  file_path = paste0("CDT-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{data.frame of dekadal climatic data in tidy format i.e. one row
per dekad (per station) and one column per element.}

\item{station}{Name of the station identifying column in \code{data}.}

\item{element}{Name of the element column in \code{data}.}

\item{date}{Name of the date column in \code{data}. Required if \code{type =
  "daily"}. If \code{type = "dekad"} this is only needed if \code{year},
\code{month}, and \code{dekad} are not specified.}

\item{latitude}{Name of the latitude column in \code{metadata}.}

\item{longitude}{Name of the longitude column in \code{metadata}.}

\item{altitude}{Name of the altitude column in \code{metadata}.}

\item{year}{Name of the year column in \code{data}. Only needed if \code{type
  = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::year(data[[date]])}.}

\item{month}{Name of the month column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::month(data[[date]])}.}

\item{dekad}{Name of the dekad column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad}
function.}

\item{metadata}{data.frame of station metadata. Use this is the station
details are in a separate data.frame with one row per station. If
specified, \code{latitude}, \code{longitude} and \code{altitude} are
assumed to be in \code{metadata} and \code{station} must be in both
\code{data} and \code{metadata} to facilitate joining.}

\item{file_path}{A character specifying the file path and file name to export}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export data in CDT format for dekad data
}
\examples{
# TODO

}
