% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_climat_messages.R
\name{export_climat_messages}
\alias{export_climat_messages}
\title{Export CLIMAT messages file(s) from daily data}
\usage{
export_climat_messages(
  data,
  date_time,
  station_id,
  year = NULL,
  month = NULL,
  mean_pressure_station = NULL,
  mean_pressure_reduced = NULL,
  mean_temp = NULL,
  mean_max_temp = NULL,
  mean_min_temp = NULL,
  mean_vapour_pressure = NULL,
  total_precip = NULL,
  total_sunshine = NULL,
  folder = getwd()
)
}
\arguments{
\item{data}{TODO}

\item{date_time}{TODO}

\item{station_id}{TODO}

\item{year}{TODO}

\item{month}{TODO}

\item{mean_pressure_station}{TODO}

\item{mean_pressure_reduced}{TODO}

\item{mean_temp}{TODO}

\item{mean_max_temp}{TODO}

\item{mean_min_temp}{TODO}

\item{mean_vapour_pressure}{TODO}

\item{total_precip}{TODO}

\item{total_sunshine}{TODO}

\item{folder}{TODO}
}
\description{
Export CLIMAT messages file(s) from daily data
}
\examples{
# TODO
}
