% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_climdex.R
\name{export_climdex}
\alias{export_climdex}
\title{Export data in the format for RClimDex}
\usage{
export_climdex(
  data,
  prcp,
  tmax,
  tmin,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  file_type = c("csv", "txt"),
  file_path = paste0("climdex-", Sys.Date()),
  ...
)
}
\arguments{
\item{data}{data.frame of daily climatic data in tidy format i.e. one row per
day and one column per element.}

\item{prcp}{Name of the precipitation/rainfall column in \code{data}.}

\item{tmax}{Name of the maximum temperature column in \code{data}.}

\item{tmin}{Name of the minimum temperature column in \code{data}.}

\item{date}{Name of the date column in \code{data}. This is only needed if
\code{year}, \code{month}, and \code{day} are not specified.}

\item{year}{Name of the year column in \code{data}. If \code{NULL} it will be
created using \code{lubridate::year(data[[date]])}.}

\item{month}{Name of the month column in \code{data}. If \code{NULL} it will
be created using \code{lubridate::month(data[[date]])}.}

\item{day}{Name of the day of the month column in \code{data}. If \code{NULL}
it will be created using \code{lubridate::day(data[[date]])}.}

\item{file_type}{A character specifying the file type to export as, either \code{"csv"} or `\code{"txt"}.}

\item{file_path}{A character specifying the file path and file name to export}

\item{...}{Other parameters passed to \code{write.table()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export data in the format for RClimDex
}
\examples{
# TODO

}
