% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim_dekad.R
\name{export_geoclim_dekad}
\alias{export_geoclim_dekad}
\title{Export dekad data in GeoCLIM format}
\usage{
export_geoclim_dekad(
  data,
  year,
  dekad,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{year}{The name of the year column in \code{data}.}

\item{dekad}{TODO}

\item{element}{The name of the column in \code{data} to apply the function to.}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{add_cols}{Names of additional metadata columns that should be included in the output}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export dekad data in GeoCLIM format
}
\examples{
# TODO
}
