% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_from_data.R
\name{get_lat_from_data}
\alias{get_lat_from_data}
\title{Get the latitude coordinate from the data}
\usage{
get_lat_from_data(datafile)
}
\arguments{
\item{datafile}{}
}
\value{

}
\description{
Get the latitude coordinate from the data
}
