% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lon_from_data.R
\name{get_lon_from_data}
\alias{get_lon_from_data}
\title{Get the longitude coordinate from the data}
\usage{
get_lon_from_data(datafile)
}
\arguments{
\item{datafile}{TODO}
}
\value{

}
\description{
Get the longitude coordinate from the data
}
\examples{
# TODO
}
