% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Title}
\usage{
histogram_plot(
  data,
  date_time,
  elements,
  station = NULL,
  facets = c("stations", "elements", "both", "none"),
  position = c("identity", "dodge", "dodge2", "stack"),
  plot_type = c("histogram", "density", "frequency"),
  binwidth = NULL,
  bins = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  breaks = NULL
)
}
\arguments{
\item{data}{}

\item{date_time}{}

\item{elements}{}

\item{station}{}

\item{facets}{}

\item{position}{}

\item{plot_type}{}

\item{binwidth}{}

\item{bins}{}

\item{na.rm}{}

\item{orientation}{}

\item{show.legend}{}

\item{breaks}{}
}
\value{

}
\description{
Title
}
