% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_iri.R
\name{import_from_iri}
\alias{import_from_iri}
\title{Import data from IRI (International Research Institute for Climate and Society)}
\usage{
import_from_iri(download_from, data_file, path, X1, X2, Y1, Y2, get_area_point)
}
\arguments{
\item{download_from}{Source(?) to download the data from. Takes values \code{"CHIRPS_V2P0"}, \code{"TAMSTAT"}, \code{"NOAA_ARC2"}, \code{"NOAA_RFE2"}, \code{"NOAA_CMORPH_DAILY"}, \code{"NASA_TRMM_3B42"}}

\item{data_file}{}

\item{path}{}

\item{X1}{}

\item{X2}{}

\item{Y1}{}

\item{Y2}{}

\item{get_area_point}{}
}
\value{

}
\description{
Import data from IRI (International Research Institute for Climate and Society)
}
