% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lon_dataframe.R
\name{lat_lon_dataframe}
\alias{lat_lon_dataframe}
\title{Get the latitude and longitude coordinates from a data frame}
\usage{
lat_lon_dataframe(datafile)
}
\arguments{
\item{datafile}{TODO}
}
\value{

}
\description{
Get the latitude and longitude coordinates from a data frame
}
\examples{
# TODO
}
