% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_nc_as_data_frame.R
\name{multiple_nc_as_data_frame}
\alias{multiple_nc_as_data_frame}
\title{Title}
\usage{
multiple_nc_as_data_frame(
  path,
  vars,
  keep_raw_time = TRUE,
  include_metadata = TRUE,
  boundary = NULL,
  lon_points = NULL,
  lat_points = NULL,
  id_points = NULL,
  show_requested_points = TRUE,
  great_circle_dist = TRUE,
  id = "id"
)
}
\arguments{
\item{path}{}

\item{vars}{}

\item{keep_raw_time}{}

\item{include_metadata}{}

\item{boundary}{}

\item{lon_points}{}

\item{lat_points}{}

\item{id_points}{}

\item{show_requested_points}{}

\item{great_circle_dist}{}

\item{id}{}
}
\value{

}
\description{
Title
}
