% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_as_data_frame.R
\name{nc_as_data_frame}
\alias{nc_as_data_frame}
\title{Title}
\usage{
nc_as_data_frame(
  nc,
  vars,
  keep_raw_time = TRUE,
  include_metadata = TRUE,
  boundary = NULL,
  lon_points = NULL,
  lat_points = NULL,
  id_points = NULL,
  show_requested_points = TRUE,
  great_circle_dist = TRUE
)
}
\arguments{
\item{nc}{TODO}

\item{vars}{TODO}

\item{keep_raw_time}{TODO}

\item{include_metadata}{TODO}

\item{boundary}{TODO}

\item{lon_points}{TODO}

\item{lat_points}{TODO}

\item{id_points}{TODO}

\item{show_requested_points}{TODO}

\item{great_circle_dist}{TODO}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
