% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_get_dim_min_max.R
\name{nc_get_dim_min_max}
\alias{nc_get_dim_min_max}
\title{Title}
\usage{
nc_get_dim_min_max(nc, dimension, time_as_date = TRUE)
}
\arguments{
\item{nc}{}

\item{dimension}{}

\item{time_as_date}{}
}
\value{

}
\description{
Title
}
