% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_CPT.R
\name{output_CPT}
\alias{output_CPT}
\title{Outputs data in the format for the CPT software}
\usage{
output_CPT(
  data,
  lat_lon_data,
  station_latlondata,
  latitude,
  longitude,
  station,
  year,
  element,
  long_data = TRUE,
  na_code = -999
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{lat_lon_data}{\code{data.frame} The name of the metadata to calculate from.}

\item{station_latlondata}{\code{character(1)} The name of the station column in \code{lat_lon_data}, or \code{data} if \code{long_data = FALSE}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{lat_lon_data}, or \code{data} if \code{long_data = FALSE}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{lat_lon_data}, or \code{data} if \code{long_data = FALSE}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{long_data}{\code{logical(1)} Whether all columns are in \code{data}. If all data is in one data frame then must have \code{long_data = TRUE}.}

\item{na_code}{\code{numeric(1)} Indicator for NA values in data.}
}
\value{
A data.frame formatted for use in CPT.
}
\description{
Outputs data in the format for the CPT software
}
\examples{
# Create summary data
yearly_niger <- daily_niger \%>\% dplyr::group_by(station_name, year) \%>\%
    dplyr::summarise(mean_rain = mean(rain))
output_CPT(data = yearly_niger, lat_lon_data = stations_niger,
           station_latlondata = "station_name", latitude = "lat", longitude = "long",
           station = "station_name", year = "year", element = "mean_rain")
}
