% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt_daily.R
\name{prepare_cdt_daily}
\alias{prepare_cdt_daily}
\title{Prepare CDT for Daily data}
\usage{
prepare_cdt_daily(
  data,
  station,
  element,
  date,
  metadata = NULL,
  latitude,
  longitude,
  altitude
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{date}{The name of the date column in \code{data}. Required if \code{type = "daily"}. If \code{type = "dekad"} this is only needed if \code{year}, \code{month}, and \code{dekad} are not specified.}

\item{metadata}{The metadata data.frame.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}
}
\value{

}
\description{
Prepare CDT for Daily data
}
\examples{
# With daily_niger data:
data("daily_niger"); data("stations_niger")
prepare_cdt_daily(data = daily_niger, date = "date", station = "station_name",
                  element = "tmax", metadata = stations_niger, 
                  latitude = "lat", longitude = "long", altitude = "alt")
}
