% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt_dekad.R
\name{prepare_cdt_dekad}
\alias{prepare_cdt_dekad}
\title{Prepare CDT for dekad data}
\usage{
prepare_cdt_dekad(
  data,
  station,
  element,
  date = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL,
  latitude,
  longitude,
  altitude
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{date}{The name of the date column in \code{data}. Required if \code{type = "daily"}. If \code{type = "dekad"} this is only needed if \code{year}, \code{month}, and \code{dekad} are not specified.}

\item{year}{The name of the year column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{dekad}{The name of the dekad column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad} function.}

\item{metadata}{The metadata data.frame.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}
}
\value{

}
\description{
Prepare CDT for dekad data
}
