% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_climdex.R
\name{prepare_climdex}
\alias{prepare_climdex}
\title{Prepare data in the format for RClimDex}
\usage{
prepare_climdex(
  data,
  prcp,
  tmax,
  tmin,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  na = -99.9
)
}
\arguments{
\item{data}{data.frame of daily climatic data in tidy format i.e. one row per
day and one column per element.}

\item{prcp}{Name of the precipitation/rainfall column in \code{data}.}

\item{tmax}{Name of the maximum temperature column in \code{data}.}

\item{tmin}{Name of the minimum temperature column in \code{data}.}

\item{date}{Name of the date column in \code{data}. This is only needed if
\code{year}, \code{month}, and \code{day} are not specified.}

\item{year}{Name of the year column in \code{data}. If \code{NULL} it will be
created using \code{lubridate::year(data[[date]])}.}

\item{month}{Name of the month column in \code{data}. If \code{NULL} it will
be created using \code{lubridate::month(data[[date]])}.}

\item{day}{Name of the day of the month column in \code{data}. If \code{NULL}
it will be created using \code{lubridate::day(data[[date]])}.}

\item{na}{The value to replace missing values with. The default (-99.9)
should be used to comply with the RClimDex format requirements.}
}
\value{
A data.frame formatted for use in RClimDex
}
\description{
Prepare data in the format for RClimDex
}
\examples{
# Preparing the daily_niger data for export to ClimDex
prepare_climdex(data = daily_niger, date = "date", prcp = "rain", tmax = "tmax", tmin = "tmin")

}
