% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim.R
\name{prepare_geoclim}
\alias{prepare_geoclim}
\title{Prepare dekad or pentad data in GeoCLIM format}
\usage{
prepare_geoclim(
  data,
  year,
  type_col,
  element,
  station_id,
  latitude,
  longitude,
  type = c("dekad", "pentad"),
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{type_col}{TODO}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{type}{TODO}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{add_cols}{Names of additional metadata columns that should be included in the output}
}
\value{

}
\description{
Prepare dekad or pentad data in GeoCLIM format
}
\examples{
# TODO
}
