% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_dekad.R
\name{prepare_geoclim_dekad}
\alias{prepare_geoclim_dekad}
\title{Prepare dekad data in GeoCLIM format}
\usage{
prepare_geoclim_dekad(
  data,
  year,
  dekad,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{year}{The name of the year column in \code{data}.}

\item{dekad}{TODO}

\item{element}{The name of the column in \code{data} to apply the function to.}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{add_cols}{TODO}
}
\value{

}
\description{
Prepare dekad data in GeoCLIM format
}
\examples{
# TODO
}
