% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_month.R
\name{prepare_geoclim_month}
\alias{prepare_geoclim_month}
\title{Prepare monthly data in GeoCLIM format}
\usage{
prepare_geoclim_month(
  data,
  year,
  month,
  element,
  station_id,
  latitude,
  longitude,
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{month}{\code{character(1)} The name of the month column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{add_cols}{TODO}
}
\value{

}
\description{
Prepare monthly data in GeoCLIM format
}
\examples{
# TODO
}
