% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_pentad.R
\name{prepare_geoclim_pentad}
\alias{prepare_geoclim_pentad}
\title{Prepare pentad data in GeoCLIM format}
\usage{
prepare_geoclim_pentad(
  data,
  year,
  pentad,
  element,
  station_id,
  latitude,
  longitude,
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{year}{The name of the year column in \code{data}.}

\item{pentad}{TODO}

\item{element}{The name of the column in \code{data} to apply the function to.}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{add_cols}{TODO}
}
\value{

}
\description{
Prepare pentad data in GeoCLIM format
}
\examples{
# TODO
}
