% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei_input.R
\name{spei_input}
\alias{spei_input}
\title{Title}
\usage{
spei_input(data, station, year, month, element)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{element}{The name of the column(s) in \code{data} to apply the condition to.}
}
\value{

}
\description{
Title
}
