% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei_output.R
\name{spei_output}
\alias{spei_output}
\title{Title}
\usage{
spei_output(x, data, station, year, month)
}
\arguments{
\item{x}{An object of class 'spei'.}

\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}
}
\value{

}
\description{
This function extracts the SPEI/SPI column from an spei object x.
}
