% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations_niger}
\alias{stations_niger}
\title{Details of eleven weather stations in Niger}
\format{
A data frame with 11 rows and 6 variables:
\describe{
\item{id}{the id of the station}
\item{station_name}{the name of the station}
\item{lat}{the latitude of the station, in degrees}
\item{long}{the longitude of the station, in degrees}
\item{alt}{the altitude of the station, in m}
\item{daily}{logical, if \code{TRUE} daily data is available for this
station in the \code{daily_niger} dataset}
}
}
\source{
Meteo-France \url{https://meteofrance.com/}
}
\usage{
stations_niger
}
\description{
A dataset containing details of eleven weather stations in Niger.
Measurements from four of these stations are available in the
\code{daily_niger} dataset.
}
\keyword{datasets}
