% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_inventory_data.r
\name{summarise_inventory_data}
\alias{summarise_inventory_data}
\title{Title}
\usage{
summarise_inventory_data(data, station, elements)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.
The summary table is calculated separately for each station.}

\item{elements}{The name of the columns in \code{data} to apply the summary calculation to.}
}
\value{

}
\description{
Title
}
\examples{
data(daily_niger)
summarise_inventory_data(data = daily_niger, station = "station_name", elements = c("tmax", "tmin"))
}
