% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Windrose from the clifro package}
\usage{
windrose(
  data,
  speed,
  direction,
  facet = NULL,
  n_directions = 12,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  ggtheme = c("grey", "gray", "bw", "linedraw", "light", "minimal", "classic"),
  legend_title = "Wind Speed",
  calm_wind = 0,
  variable_wind = 990,
  n_col = NULL
)
}
\arguments{
\item{data}{}

\item{speed}{}

\item{direction}{}

\item{facet}{}

\item{n_directions}{}

\item{n_speeds}{}

\item{speed_cuts}{}

\item{col_pal}{}

\item{ggtheme}{}

\item{legend_title}{}

\item{calm_wind}{}

\item{variable_wind}{}

\item{n_col}{}
}
\value{

}
\description{
A wrapper for plotting a windrose of wind speed and direction using ggplot2
}
