% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwr_export.R
\name{wwr_export}
\alias{wwr_export}
\title{Title}
\usage{
wwr_export(
  data,
  year,
  month,
  mean_station_pressure,
  mean_sea_level_pressure,
  mean_temp,
  total_precip,
  mean_max_temp,
  mean_min_temp,
  mean_rel_hum,
  link,
  link_by,
  station_data,
  wmo_number,
  latitude,
  longitude,
  country_name,
  station_name,
  height_station,
  height_barometer,
  wigos_identifier,
  folder
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{mean_station_pressure}{TODO}

\item{mean_sea_level_pressure}{TODO}

\item{mean_temp}{TODO}

\item{total_precip}{TODO}

\item{mean_max_temp}{TODO}

\item{mean_min_temp}{TODO}

\item{mean_rel_hum}{TODO}

\item{link}{TODO}

\item{link_by}{TODO}

\item{station_data}{TODO}

\item{wmo_number}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{country_name}{TODO}

\item{station_name}{TODO}

\item{height_station}{TODO}

\item{height_barometer}{TODO}

\item{wigos_identifier}{TODO}

\item{folder}{TODO}
}
\value{

}
\description{
Reshape data into formats required by WMO for submission of climatic data
this gives Yearly data records with monthly and annual data for a particular year.
}
\examples{
# TODO
}
