% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_missing.R
\name{climatic_missing}
\alias{climatic_missing}
\title{Summarise missing data in a data frame}
\usage{
climatic_missing(
  data,
  date_time,
  elements,
  station_id = NULL,
  start = TRUE,
  end = FALSE
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{The name of the date column in \code{data}.}

\item{elements}{The name of the column in \code{data} to apply the function to.}

\item{station_id}{The name of the station column in \code{data}, if the data are for multiple station.
The calculations are performed separately for each station.}

\item{start}{\code{logical(1)} If \code{TRUE} start date as ...}

\item{end}{\code{logical(1)} If \code{TRUE} set end date as ...}
}
\value{
Data frame summarising the missing data
}
\description{
\code{climatic_missing} returns a data table displaying the number and
percentage of missing values for an element (and station) in a given time period.
The total number of full years are also given.
}
\examples{
# Summarising missing data in the rainfall, temperature, and sun columns
climatic_missing(data = daily_niger, date_time = "date", 
                 elements = c("rain", "tmax", "tmin", "sunh"),
                 station_id = "station_name")

}
