% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_summary.R
\name{climatic_summary}
\alias{climatic_summary}
\title{Title}
\usage{
climatic_summary(
  data,
  date_time,
  station = NULL,
  elements,
  year = NULL,
  month = NULL,
  dekad = NULL,
  pentad = NULL,
  to = c("hourly", "daily", "pentad", "dekadal", "monthly", "annual-within-year",
    "annual", "longterm-monthly", "longterm-within-year", "station", "overall"),
  by = NULL,
  doy = NULL,
  doy_first = 1,
  doy_last = 366,
  summaries = c(n = "dplyr::n"),
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL,
  first_date = FALSE,
  n_dates = FALSE,
  last_date = FALSE,
  summaries_params = list(),
  names = "{.fn}_{.col}"
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{date_time}{The name of the date column in \code{data}.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{elements}{The name of the column in \code{data} to apply the spell condition to}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{dekad}{TODO}

\item{pentad}{TODO}

\item{to}{TODO}

\item{by}{TODO}

\item{doy}{The name of the day of the year (1-366) column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[doy]])}.}

\item{doy_first}{The first day of the year.}

\item{doy_last}{The last day of the year}

\item{summaries}{A named character vector of summary functions. The names are
the used as the column names in the results. The values can be any function
name as a string. e.g. c(mean = "mean", st_dv = "sd", n_na =
"naflex::na_n")}

\item{na_rm}{If \code{TRUE} all \code{na_} parameters are ignored and missing
values are removed. If \code{FALSE} missing values are not removed unless
any \code{na_} parameters are specified.}

\item{na_prop}{Max proportion of missing values allowed}

\item{na_n}{Max number of missing values allowed}

\item{na_consec}{Max number of consecutive missing values allowed}

\item{na_n_non}{Min number of non-missing values required}

\item{first_date}{If \code{TRUE} the first instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. first \code{date_time} when the maximum occurred.}

\item{n_dates}{If \code{TRUE} the number of \code{date_time} points when the value equals the summary value is included. Generally only used for extreme summaries i.e. number of days in which the minimum occurred.}

\item{last_date}{If \code{TRUE} the last instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. last \code{date_time} when the maximum occurred.}

\item{summaries_params}{Additional parameters to pass to \code{summaries}.
Must be a list of lists with the same names as \code{summaries}.}

\item{names}{Format of column names. Passed to \code{.names} in
\code{dplyr::across}}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
