% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.R
\name{climdex}
\alias{climdex}
\title{Title}
\usage{
climdex(
  data,
  station,
  date,
  year,
  month,
  prec,
  tmax,
  tmin,
  indices,
  freq = "annual",
  base.range = c(1961, 1990),
  n = 5,
  northern.hemisphere = TRUE,
  quantiles = NULL,
  temp.qtiles = c(0.1, 0.9),
  prec.qtiles = c(0.95, 0.99),
  max.missing.days = c(annual = 15, monthly = 3),
  min.base.data.fraction.present = 0.1,
  spells.can.span.years = FALSE,
  gsl.mode = c("GSL", "GSL_first", "GSL_max", "GSL_sum"),
  threshold = 1
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{date}{The name of the date column in \code{data}.}

\item{year}{The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{prec}{The name of the precipitation column in \code{data}.}

\item{tmax}{The name of the maximum temperature column in \code{data}.}

\item{tmin}{The name of the minimum temperature column in \code{data}.}

\item{indices}{TODO}

\item{freq}{TODO}

\item{base.range}{TODO}

\item{n}{TODO}

\item{northern.hemisphere}{TODO}

\item{quantiles}{TODO}

\item{temp.qtiles}{TODO}

\item{prec.qtiles}{TODO}

\item{max.missing.days}{TODO}

\item{min.base.data.fraction.present}{TODO}

\item{spells.can.span.years}{TODO}

\item{gsl.mode}{TODO}

\item{threshold}{TODO}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
