% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex_single_station.R
\name{climdex_single_station}
\alias{climdex_single_station}
\title{Title}
\usage{
climdex_single_station(
  ci,
  indices,
  year,
  month,
  freq = "annual",
  spells.can.span.years = FALSE,
  gsl.mode = c("GSL", "GSL_first", "GSL_max", "GSL_sum"),
  threshold = 1
)
}
\arguments{
\item{ci}{TODO}

\item{indices}{TODO}

\item{year}{TODO}

\item{month}{TODO}

\item{freq}{TODO}

\item{spells.can.span.years}{TODO}

\item{gsl.mode}{TODO}

\item{threshold}{TODO}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
