% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dec_deg.R
\name{convert_to_dec_deg}
\alias{convert_to_dec_deg}
\title{Title}
\usage{
convert_to_dec_deg(dd, mm = 0, ss = 0, dir)
}
\arguments{
\item{dd}{}

\item{mm}{}

\item{ss}{}

\item{dir}{Direction to --- to. Takes values \code{"N"}, \code{"E"}, \code{"S"}, or \code{"W"}}
}
\value{

}
\description{
Title
}
