% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdt.R
\name{export_cdt}
\alias{export_cdt}
\title{Export data in CDT format}
\usage{
export_cdt(
  data,
  date,
  year,
  station,
  element,
  metadata = NULL,
  latitude,
  longitude,
  altitude,
  file_path = paste0("CDT-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date}{The name of the date column in \code{data}.}

\item{year}{The name of the year column in \code{data}.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{metadata}{The metadata data.frame.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export data in CDT format
}
\examples{
# TODO



}
