% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdt_daily.R
\name{export_cdt_daily}
\alias{export_cdt_daily}
\title{Export daily data in the format for CDT}
\usage{
export_cdt_daily(
  data,
  station,
  element,
  date_time,
  latitude,
  longitude,
  altitude,
  metadata = NULL,
  file_path = paste0("CDT-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{data.frame of daily climatic data in tidy format i.e. one row
per day (per station) and one column per element.}

\item{station}{Name of the station identifying column in \code{data}.}

\item{element}{Name of the element column in \code{data}.}

\item{date_time}{Name of the date column in \code{data}. Required if \code{type =
  "daily"}. If \code{type = "dekad"} this is only needed if \code{year},
\code{month}, and \code{dekad} are not specified.}

\item{latitude}{Name of the latitude column in \code{metadata}.}

\item{longitude}{Name of the longitude column in \code{metadata}.}

\item{altitude}{Name of the altitude column in \code{metadata}.}

\item{metadata}{data.frame of station metadata. Use this is the station
details are in a separate data.frame with one row per station. If
specified, \code{latitude}, \code{longitude} and \code{altitude} are
assumed to be in \code{metadata} and \code{station} must be in both
\code{data} and \code{metadata} to facilitate joining.}

\item{file_path}{\code{character(1)} A character specifying the file path and file name to export.}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export daily data in the format for CDT
}
\examples{
# To write daily_niger data to CDT format
## NOT RUN ##
#export_cdt_daily(data = daily_niger, station = "station_name", element = "rain", type = "daily",
#           date_time = "date", latitude = "lat", longitude = "long", altitude = "alt",
#           metadata = stations_niger)
}
