% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdt_daily.R
\name{export_cdt_daily}
\alias{export_cdt_daily}
\title{Export data in CDT format for daily data}
\usage{
export_cdt_daily(
  data,
  station,
  element,
  date,
  metadata = NULL,
  latitude,
  longitude,
  altitude,
  file_path = paste0("CDT-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{date}{The name of the date column in \code{data}. Required if \code{type = "daily"}. If \code{type = "dekad"} this is only needed if \code{year}, \code{month}, and \code{dekad} are not specified.}

\item{metadata}{The metadata data.frame.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export data in CDT format for daily data
}
\examples{
# TODO



}
