% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cdt_dekad.R
\name{export_cdt_dekad}
\alias{export_cdt_dekad}
\title{Export data in CDT format for dekad data}
\usage{
export_cdt_dekad(
  data,
  station,
  element,
  date,
  latitude,
  longitude,
  altitude,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL,
  file_path = paste0("CDT-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{date}{The name of the date column in \code{data}. Required if \code{type = "daily"}. If \code{type = "dekad"} this is only needed if \code{year}, \code{month}, and \code{dekad} are not specified.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}

\item{year}{The name of the year column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{dekad}{The name of the dekad column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad} function.}

\item{metadata}{The metadata data.frame.}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export data in CDT format for dekad data
}
\examples{
# TODO

}
