% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_climdex.R
\name{export_climdex}
\alias{export_climdex}
\title{Export ClimDex data}
\usage{
export_climdex(
  data,
  prcp,
  tmax,
  tmin,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  file_type = c("csv", "txt"),
  file_path = paste0("climdex-", Sys.Date()),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{prcp}{The name of the rainfall column in \code{data}.}

\item{tmax}{The name of the maximum temperature column in \code{data}.}

\item{tmin}{The name of the minimum temperature column in \code{data}.}

\item{date}{The name of the date column in \code{data}. This is only needed if \code{year}, \code{month}, and \code{day} are not specified.}

\item{year}{The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{day}{The name of the day column in \code{data}. If \code{NULL} it will be created using \code{lubridate::day(data[[date]])}.}

\item{file_type}{Whether to convert to \code{csv} or `\code{txt} format.}

\item{file_path}{Default name of the file}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export ClimDex data
}
\examples{
# TODO

}
