% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim.R
\name{export_geoclim}
\alias{export_geoclim}
\title{Export dekad or pentad data in GeoCLIM format}
\usage{
export_geoclim(
  data,
  year,
  type = c("dekad", "pentad"),
  type_col,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{}

\item{year}{}

\item{type}{}

\item{type_col}{}

\item{element}{}

\item{metadata}{}

\item{join_by}{}

\item{station_id}{}

\item{latitude}{}

\item{longitude}{}

\item{add_cols}{}

\item{file_path}{}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export dekad or pentad data in GeoCLIM format
}
