% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim.R
\name{export_geoclim}
\alias{export_geoclim}
\title{Export dekad or pentad data in GeoCLIM format}
\usage{
export_geoclim(
  data,
  year,
  type_col,
  element,
  station_id,
  latitude,
  longitude,
  type = c("dekad", "pentad"),
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{year}{The name of the year column in \code{data}.}

\item{type_col}{TODO}

\item{element}{The name of the column in \code{data} to apply the function to.}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{type}{TODO}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{add_cols}{TODO}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export dekad or pentad data in GeoCLIM format
}
\examples{
# TODO
}
