% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim.R
\name{export_geoclim}
\alias{export_geoclim}
\title{Export dekad or pentad data in GeoCLIM format}
\usage{
export_geoclim(
  data,
  year,
  type_col,
  element,
  station_id,
  latitude,
  longitude,
  type = c("dekad", "pentad"),
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{type_col}{\code{character(1)} The name of the dekad or pentad column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{\code{character(1)} The name of the station column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{type}{\code{character(1)} Whether the data is in \code{dekad} or \code{pentad} format.}

\item{metadata}{\code{data.frame} The metadata data.frame to calculate from.}

\item{join_by}{\code{character} The variable(s) to merge the \code{data} and \code{metadata} data frames.}

\item{add_cols}{Names of additional metadata columns that should be included in the output}

\item{file_path}{\code{character(1)} A character specifying the file path and file name to export.}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data.
}
\description{
Export dekad or pentad data in GeoCLIM format
}
\examples{
# TODO
}
