% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim_month.R
\name{export_geoclim_month}
\alias{export_geoclim_month}
\title{Export monthly data in GeoCLIM format}
\usage{
export_geoclim_month(
  data,
  year,
  month,
  element,
  station_id,
  latitude,
  longitude,
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{month}{The name of the month column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{metadata}{TODO}

\item{join_by}{TODO}

\item{add_cols}{TODO}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export monthly data in GeoCLIM format
}
\examples{
# TODO
}
