% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim_pentad.R
\name{export_geoclim_pentad}
\alias{export_geoclim_pentad}
\title{Export pentad data in GeoCLIM format}
\usage{
export_geoclim_pentad(
  data,
  year,
  pentad,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{pentad}{\code{character(1)} The name of the pentad column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{metadata}{\code{data.frame} The metadata data.frame to calculate from.}

\item{join_by}{\code{character} The variable(s) to merge the \code{data} and \code{metadata} data frames.}

\item{station_id}{\code{character(1)} The name of the station column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{add_cols}{\code{character} Names of additional metadata columns that should be included in the output}

\item{file_path}{\code{character(1)} A character specifying the file path and file name to export.}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data.
}
\description{
\code{export_geoclim_pentad} takes a data frame that is given by
by pentad. This data frame is then rearranged using \code{prepare_geoclim_pentad} to a
format suitable for use in GeoCLIM, and then written to a file or connection.
}
\examples{
# Calculate pentad summaries for the rainfall column
pentad_data <- daily_niger \%>\% dplyr::mutate(pentad = pentad(date))
# Summarise the data
summary_data <- pentad_data \%>\% dplyr::group_by(station_name, year, pentad) \%>\%
      dplyr::summarise(mean_rain = mean(rain, na.rm = TRUE))
# NOT RUN: Export the data to CSV format
#export_geoclim_pentad(data = summary_data, year = "year",
#                     station_id = "station_name",
#                     pentad = "pentad",
#                     element = "mean_rain", metadata = stations_niger, 
#                     join_by = "station_name",
#                     latitude = "lat", longitude = "long")
}
