% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geoclim_pentad.R
\name{export_geoclim_pentad}
\alias{export_geoclim_pentad}
\title{Export pentad data in GeoCLIM format}
\usage{
export_geoclim_pentad(
  data,
  year,
  pentad,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL,
  file_path = paste0("GEOCLIM-", element, ".csv"),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{pentad}{\code{character(1)} The name of the pentad column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{metadata}{\code{data.frame} The metadata data.frame to calculate from.}

\item{join_by}{\code{character} The variable(s) to merge the \code{data} and \code{metadata} data frames.}

\item{station_id}{\code{character(1)} The name of the station column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{add_cols}{Names of additional metadata columns that should be included in the output}

\item{file_path}{TODO}

\item{...}{Other parameters passed to \code{write.csv()}}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Export pentad data in GeoCLIM format
}
\examples{
# TODO
}
