% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_plot.R
\name{inventory_plot}
\alias{inventory_plot}
\title{Inventory Plot}
\usage{
inventory_plot(
  data,
  date,
  elements,
  station = NULL,
  year = NULL,
  doy = NULL,
  year_doy_plot = FALSE,
  facet_by = NULL,
  facet_x_size = 7,
  facet_y_size = 11,
  title = "Inventory Plot",
  plot.title_size = NULL,
  plot.title_hjust = 0.5,
  x_title = NULL,
  y_title = NULL,
  x_scale_from = NULL,
  x_scale_to = NULL,
  x_scale_by = NULL,
  y_date_format = NULL,
  y_date_scale_by = NULL,
  y_date_scale_step = 1,
  facet_scales = "fixed",
  facet_dir = "h",
  facet_x_margin = ggplot2::margin(1, 0, 1, 0),
  facet_y_margin = ggplot2::margin(1, 0, 1, 0),
  facet_nrow = NULL,
  facet_ncol = NULL,
  missing_colour = "red",
  present_colour = "grey",
  missing_label = "Missing",
  present_label = "Present",
  display_rain_days = FALSE,
  rain = NULL,
  rain_cats = list(breaks = c(0, 0.85, Inf), labels = c("Dry", "Rain"), key_colours =
    c("tan3", "blue")),
  coord_flip = FALSE
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date}{The name of the date column in \code{data}.}

\item{elements}{The name of the column in \code{data} to apply the function to.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{doy}{The name of the day of the year (1-366) column in \code{data}.
If \code{doy} is \code{NULL} then it can be calculated as \code{yday_366(data[[date]])} if \code{date} is provided.}

\item{year_doy_plot}{TODO}

\item{facet_by}{TODO}

\item{facet_x_size}{TODO}

\item{facet_y_size}{TODO}

\item{title}{The text for the title.}

\item{plot.title_size}{TODO}

\item{plot.title_hjust}{TODO}

\item{x_title}{The text for the x-axis.}

\item{y_title}{The text for the y-axis.}

\item{x_scale_from}{TODO - see ?scale_continuous}

\item{x_scale_to}{TODO}

\item{x_scale_by}{TODO}

\item{y_date_format}{TODO}

\item{y_date_scale_by}{TODO}

\item{y_date_scale_step}{TODO}

\item{facet_scales}{TODO}

\item{facet_dir}{TODO}

\item{facet_x_margin}{TODO}

\item{facet_y_margin}{TODO}

\item{facet_nrow}{TODO}

\item{facet_ncol}{TODO}

\item{missing_colour}{TODO}

\item{present_colour}{TODO}

\item{missing_label}{TODO}

\item{present_label}{TODO}

\item{display_rain_days}{TODO}

\item{rain}{TODO}

\item{rain_cats}{TODO}

\item{coord_flip}{TODO}
}
\value{

}
\description{
Produces an inventory of available and missing data
}
\examples{
# TODO
}
