% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_plot.R
\name{inventory_plot}
\alias{inventory_plot}
\title{Inventory Plot}
\usage{
inventory_plot(
  data,
  date,
  elements,
  station = NULL,
  year = NULL,
  doy = NULL,
  year_doy_plot = FALSE,
  coord_flip = FALSE,
  facet_by = NULL,
  facet_xsize = 7,
  facet_ysize = 11,
  title = "Inventory Plot",
  subtitle = NULL,
  caption = NULL,
  title_size = NULL,
  subtitle_size = NULL,
  caption_size = NULL,
  xaxis_title = NULL,
  xaxis_title_size = NULL,
  yaxis_title = NULL,
  yaxis_title_size = NULL,
  xaxis_text_size = NULL,
  xaxis_text_angle = NULL,
  yaxis_text_size = NULL,
  yaxis_text_angle = NULL,
  xaxis_scale_from = NULL,
  xaxis_scale_to = NULL,
  xaxis_scale_by = NULL,
  yaxis_date_format = NULL,
  yaxis_date_scale_by = NULL,
  yaxis_date_scale_step = 1,
  legend_position = NULL,
  facet_scales = "fixed",
  facet_dir = "h",
  facet_nrow = NULL,
  facet_ncol = NULL,
  missing_colour = "red",
  present_colour = "grey",
  missing_label = "Missing",
  present_label = "Present",
  display_rain_days = FALSE,
  rain = NULL,
  rain_cats = list(breaks = c(0, 0.85, Inf), labels = c("Dry", "Rain"), key_colours =
    c("tan3", "blue"))
)
}
\arguments{
\item{data}{}

\item{date}{}

\item{elements}{}

\item{station}{}

\item{year}{}

\item{doy}{}

\item{year_doy_plot}{}

\item{coord_flip}{}

\item{facet_by}{}

\item{facet_xsize}{}

\item{facet_ysize}{}

\item{title}{}

\item{subtitle}{}

\item{caption}{}

\item{title_size}{}

\item{subtitle_size}{}

\item{caption_size}{}

\item{xaxis_title}{}

\item{xaxis_title_size}{}

\item{yaxis_title}{}

\item{yaxis_title_size}{}

\item{xaxis_text_size}{}

\item{xaxis_text_angle}{}

\item{yaxis_text_size}{}

\item{yaxis_text_angle}{}

\item{xaxis_scale_from}{}

\item{xaxis_scale_to}{}

\item{xaxis_scale_by}{}

\item{yaxis_date_format}{}

\item{yaxis_date_scale_by}{}

\item{yaxis_date_scale_step}{}

\item{legend_position}{}

\item{facet_scales}{}

\item{facet_dir}{}

\item{facet_nrow}{}

\item{facet_ncol}{}

\item{missing_colour}{}

\item{present_colour}{}

\item{missing_label}{}

\item{present_label}{}

\item{display_rain_days}{}

\item{rain}{}

\item{rain_cats}{}
}
\value{

}
\description{
Produces an inventory of available and missing data
}
