% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_plot.R
\name{inventory_plot}
\alias{inventory_plot}
\title{Inventory Plot}
\usage{
inventory_plot(
  data,
  date,
  elements,
  station = NULL,
  year = NULL,
  doy = NULL,
  year_doy_plot = FALSE,
  facet_by = NULL,
  facet_x_size = 7,
  facet_y_size = 11,
  title = "Inventory Plot",
  plot.title_size = NULL,
  plot.title_hjust = 0.5,
  x_title = NULL,
  y_title = NULL,
  x_scale_from = NULL,
  x_scale_to = NULL,
  x_scale_by = NULL,
  y_date_format = NULL,
  y_date_scale_by = NULL,
  y_date_scale_step = 1,
  facet_scales = "fixed",
  facet_dir = "h",
  facet_x_margin = ggplot2::margin(1, 0, 1, 0),
  facet_y_margin = ggplot2::margin(1, 0, 1, 0),
  facet_nrow = NULL,
  facet_ncol = NULL,
  missing_colour = "red",
  present_colour = "grey",
  missing_label = "Missing",
  present_label = "Present",
  display_rain_days = FALSE,
  rain = NULL,
  rain_cats = list(breaks = c(0, 0.85, Inf), labels = c("Dry", "Rain"), key_colours =
    c("tan3", "blue")),
  coord_flip = FALSE
)
}
\arguments{
\item{data}{}

\item{date}{}

\item{elements}{}

\item{station}{}

\item{year}{}

\item{doy}{}

\item{year_doy_plot}{}

\item{facet_by}{}

\item{facet_x_size}{}

\item{facet_y_size}{}

\item{title}{}

\item{plot.title_size}{}

\item{x_title}{}

\item{y_title}{}

\item{x_scale_from}{}

\item{x_scale_to}{}

\item{x_scale_by}{}

\item{y_date_format}{}

\item{y_date_scale_by}{}

\item{y_date_scale_step}{}

\item{facet_scales}{}

\item{facet_dir}{}

\item{facet_nrow}{}

\item{facet_ncol}{}

\item{missing_colour}{}

\item{present_colour}{}

\item{missing_label}{}

\item{present_label}{}

\item{display_rain_days}{}

\item{rain}{}

\item{rain_cats}{}

\item{coord_flip}{}
}
\value{

}
\description{
Produces an inventory of available and missing data
}
