% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_table.R
\name{inventory_table}
\alias{inventory_table}
\title{Inventory Table}
\usage{
inventory_table(
  data,
  date,
  elements,
  station = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  missing_indicator = "M",
  observed_indicator = "X"
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{date}{The name of the date column in \code{data}. This is only needed if \code{year}, \code{month}, and \code{day} are not specified.}

\item{elements}{The name of the columns in \code{data} to apply the inventory calculation to.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.
The inventory table is calculated separately for each station.}

\item{year}{The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{day}{The name of the day column in \code{data}. If \code{NULL} it will be created using \code{lubridate::day(data[[date]])}.}

\item{missing_indicator}{Indicator to give if the data is missing. Default \code{"M"}.}

\item{observed_indicator}{Indicator to give if the data is observed. Default \code{"X"}.}
}
\value{

}
\description{
Inventory Table
}
\examples{
# TODO
inventory_table(data = niger_daily_date, station = "Name", date = "date", elements = "tmax")
}
