% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_table.R
\name{inventory_table}
\alias{inventory_table}
\title{Inventory Table}
\usage{
inventory_table(
  data,
  date_time,
  elements,
  station = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  missing_indicator = "M",
  observed_indicator = "X"
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{elements}{\code{character} The name of the elements column in \code{data} to apply the function to.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.
The inventory table is calculated separately for each station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{month}{\code{character(1)} The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date_time]])}.}

\item{day}{\code{character(1)} The name of the day column in \code{data}. If \code{NULL} it will be created using \code{lubridate::day(data[[date_time]])}.}

\item{missing_indicator}{\code{character(1)} Indicator to give if the data is missing. Default \code{"M"}.}

\item{observed_indicator}{\code{character(1)} Indicator to give if the data is observed. Default \code{"X"}.}
}
\value{

}
\description{
Inventory Table
}
\examples{
# TODO
}
