% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_as_data_frame.R
\name{nc_as_data_frame}
\alias{nc_as_data_frame}
\title{Title}
\usage{
nc_as_data_frame(
  nc,
  vars,
  keep_raw_time = TRUE,
  include_metadata = TRUE,
  boundary = NULL,
  lon_points = NULL,
  lat_points = NULL,
  id_points = NULL,
  show_requested_points = TRUE,
  great_circle_dist = TRUE
)
}
\arguments{
\item{nc}{}

\item{vars}{}

\item{keep_raw_time}{}

\item{include_metadata}{}

\item{boundary}{}

\item{lon_points}{}

\item{lat_points}{}

\item{id_points}{}

\item{show_requested_points}{}

\item{great_circle_dist}{}
}
\value{

}
\description{
Title
}
