% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_CPT.R
\name{output_CPT}
\alias{output_CPT}
\title{Title}
\usage{
output_CPT(
  data,
  lat_lon_data,
  station_latlondata,
  latitude,
  longitude,
  station,
  year,
  element,
  long.data = TRUE,
  na_code = -999
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{lat_lon_data}{TODO}

\item{station_latlondata}{TODO}

\item{latitude}{TODO}

\item{longitude}{TODO}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{long.data}{TODO}

\item{na_code}{TODO}
}
\value{

}
\description{
Title
}
\examples{
# TODO
}
