% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_CPT.R
\name{output_CPT}
\alias{output_CPT}
\title{Title}
\usage{
output_CPT(
  data,
  lat_lon_data,
  station_latlondata,
  latitude,
  longitude,
  station,
  year,
  element,
  long.data = TRUE,
  na_code = -999
)
}
\arguments{
\item{data}{}

\item{lat_lon_data}{}

\item{station_latlondata}{}

\item{latitude}{}

\item{longitude}{}

\item{station}{}

\item{year}{}

\item{element}{}

\item{long.data}{}

\item{na_code}{}
}
\value{

}
\description{
Title
}
