% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inventory_data.R
\name{plot_inventory_data}
\alias{plot_inventory_data}
\title{Title}
\usage{
plot_inventory_data(
  data,
  elements,
  station = NULL,
  y = c("count", "percentage"),
  columns = c("elements", "stations"),
  fill = c("none", "stations", "elements"),
  facets = c("stations", "elements", "none"),
  display_perc = FALSE,
  display_count = FALSE
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{elements}{The name of the columns in \code{data} to apply the summary calculation to.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.
The summary table is calculated separately for each station.}

\item{y}{Whether to display counts or percentages.}

\item{columns}{Whether to display stations or elements on the columns.}

\item{fill}{Whether to fill by stations, elements, or neither.}

\item{facets}{Whether to facet by stations, elements, or neither}

\item{display_perc}{logical. Whether to display percentage values on the plot.}

\item{display_count}{logical. Whether to display the count values on the plot.}
}
\value{

}
\description{
Title
}
\examples{
data(daily_niger) 
plot_inventory_data(data = daily_niger, station = "station_name",
                    elements = c("tmin", "tmax"),
                    columns = "stations", facets = "elements")
                    
# can omit station argument
plot_inventory_data(data = daily_niger, 
                    elements = c("tmin", "tmax"))

# can view percentage instead
plot_inventory_data(data = daily_niger, y = "percentage",
                    elements = c("tmin", "tmax"), display_perc = TRUE,
                    display_count = TRUE)
}
