% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt.R
\name{prepare_cdt}
\alias{prepare_cdt}
\title{Prepare data in format for CDT}
\usage{
prepare_cdt(
  data,
  station,
  element,
  latitude,
  longitude,
  altitude,
  type = c("dekad", "daily"),
  date = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL
)
}
\arguments{
\item{data}{data.frame of daily or dekadal climatic data in tidy format i.e.
one row per time point (per station) and one column per element.}

\item{station}{Name of the station identifying column in \code{data}.}

\item{element}{Name of the element column in \code{data}.}

\item{latitude}{Name of the latitude column in \code{metadata}.}

\item{longitude}{Name of the longitude column in \code{metadata}.}

\item{altitude}{Name of the altitude column in \code{metadata}.}

\item{type}{Character indicating the type of data, either \code{"dekad"} or
\code{"daily"}.}

\item{date}{Name of the date column in \code{data}. Required if \code{type =
  "daily"}. If \code{type = "dekad"} this is only needed if \code{year},
\code{month}, and \code{dekad} are not specified.}

\item{year}{Name of the year column in \code{data}. Only needed if \code{type
  = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::year(data[[date]])}.}

\item{month}{Name of the month column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::month(data[[date]])}.}

\item{dekad}{Name of the dekad column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad}
function.}

\item{metadata}{data.frame of station metadata. Use this is the station
details are in a separate data.frame with one row per station. If
specified, \code{latitude}, \code{longitude} and \code{altitude} are
assumed to be in \code{metadata} and \code{station} must be in both
\code{data} and \code{metadata} to facilitate joining.}
}
\value{
A data.frame formatted for use in CDT
}
\description{
Prepare data in format for CDT
}
\examples{
# create summary data
data("daily_niger"); data("stations_niger")
summary_data <- daily_niger \%>\%
dplyr::mutate(dekad_date = dekad(daily_niger$date)) \%>\%
dplyr::group_by(station_name, year, dekad_date) \%>\%
dplyr::summarise(date = dplyr::first(date), sum = sum(tmax))

prepare_cdt(data = summary_data, date = "date", year = "year",
            station = "station_name",
            element = "sum", metadata = stations_niger, 
            latitude = "lat", longitude = "long", altitude = "alt")
}
