% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt.R
\name{prepare_cdt}
\alias{prepare_cdt}
\title{Prepare CDT}
\usage{
prepare_cdt(
  data,
  date,
  year,
  element,
  metadata = NULL,
  station,
  latitude,
  longitude,
  altitude
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date}{The name of the date column in \code{data}.}

\item{year}{The name of the year column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{metadata}{The metadata data.frame.}

\item{station}{The name of the station column in \code{data}.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}
}
\value{

}
\description{
Prepare CDT
}
\examples{
# create summary data
data("daily_niger"); data("stations_niger")
summary_data <- daily_niger \%>\%
dplyr::mutate(dekad_date = dekad(daily_niger$date)) \%>\%
dplyr::group_by(station_name, year, dekad_date) \%>\%
dplyr::summarise(date = dplyr::first(date), sum = sum(tmax))

prepare_cdt(data = summary_data, date = "date", year = "year",
            station = "station_name",
            element = "sum", metadata = stations_niger, 
            latitude = "lat", longitude = "long", altitude = "alt")
}
