% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt.R
\name{prepare_cdt}
\alias{prepare_cdt}
\title{Prepare CDT}
\usage{
prepare_cdt(
  data,
  station,
  element,
  latitude,
  longitude,
  altitude,
  type = c("dekad", "daily"),
  date = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{station}{The name of the station column in \code{data}.}

\item{element}{The name of the element column in \code{data}.}

\item{latitude}{The name of the latitude column in \code{metadata}.}

\item{longitude}{The name of the longitude column in \code{metadata}.}

\item{altitude}{The name of the altitude column in \code{metadata}.}

\item{type}{Whether \code{dekad} or \code{daily} data is used.}

\item{date}{The name of the date column in \code{data}. Required if \code{type = "daily"}. If \code{type = "dekad"} this is only needed if \code{year}, \code{month}, and \code{dekad} are not specified.}

\item{year}{The name of the year column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{dekad}{The name of the dekad column in \code{data}. Only needed if \code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad} function.}

\item{metadata}{The metadata data.frame.}
}
\value{

}
\description{
Prepare CDT
}
\examples{
# create summary data
data("daily_niger"); data("stations_niger")
summary_data <- daily_niger \%>\%
dplyr::mutate(dekad_date = dekad(daily_niger$date)) \%>\%
dplyr::group_by(station_name, year, dekad_date) \%>\%
dplyr::summarise(date = dplyr::first(date), sum = sum(tmax))

prepare_cdt(data = summary_data, date = "date", year = "year",
            station = "station_name",
            element = "sum", metadata = stations_niger, 
            latitude = "lat", longitude = "long", altitude = "alt")
}
