% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt.R
\name{prepare_cdt}
\alias{prepare_cdt}
\title{Prepare data in format for CDT}
\usage{
prepare_cdt(
  data,
  station,
  element,
  latitude,
  longitude,
  altitude,
  type = c("dekad", "daily"),
  date_time = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL
)
}
\arguments{
\item{data}{\code{data.frame} data.frame of daily or dekadal climatic data in tidy format i.e.
one row per time point (per station) and one column per element.}

\item{station}{\code{character(1)} Name of the station identifying column in \code{data}.}

\item{element}{\code{character(1)} Name of the element column in \code{data}.}

\item{latitude}{\code{character(1)} Name of the latitude column in \code{metadata}.}

\item{longitude}{\code{character(1)} Name of the longitude column in \code{metadata}.}

\item{altitude}{\code{character(1)} Name of the altitude column in \code{metadata}.}

\item{type}{\code{character(1)} Character indicating the type of data, either \code{"dekad"} or
\code{"daily"}.}

\item{date_time}{\code{character(1)} Name of the date column in \code{data}. Required if \code{type =
  "daily"}. If \code{type = "dekad"} this is only needed if \code{year},
\code{month}, and \code{dekad} are not specified.}

\item{year}{\code{character(1)} Name of the year column in \code{data}. Only needed if \code{type
  = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::year(data[[date_time]])}.}

\item{month}{\code{character(1)} Name of the month column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::month(data[[date_time]])}.}

\item{dekad}{\code{character(1)} Name of the dekad column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad}
function.}

\item{metadata}{\code{data.frame} data.frame of station metadata. Use this is the station
details are in a separate data.frame with one row per station. If
specified, \code{latitude}, \code{longitude} and \code{altitude} are
assumed to be in \code{metadata} and \code{station} must be in both
\code{data} and \code{metadata} to facilitate joining.}
}
\value{
A data.frame formatted for use in CDT
}
\description{
Prepare data in format for CDT
}
\examples{
# Create summary daily data
data("daily_niger"); data("stations_niger")
cdt_data <- prepare_cdt(data = daily_niger, station = "station_name", element = "rain",
                        type = "daily", date_time = "date", latitude = "lat",
                        longitude = "long", altitude = "alt", metadata = stations_niger)
}
