% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cdt_dekad.R
\name{prepare_cdt_dekad}
\alias{prepare_cdt_dekad}
\title{Prepare data in format for CDT}
\usage{
prepare_cdt_dekad(
  data,
  station,
  element,
  date_time = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  metadata = NULL,
  latitude,
  longitude,
  altitude
)
}
\arguments{
\item{data}{\code{data.frame} data.frame of dekadal climatic data in tidy format i.e. one row
per dekad (per station) and one column per element.}

\item{station}{\code{character(1)} Name of the station identifying column in \code{data}.}

\item{element}{\code{character(1)} Name of the element column in \code{data}.}

\item{date_time}{\code{character(1)} Name of the date column in \code{data}. Required if \code{type =
  "daily"}. If \code{type = "dekad"} this is only needed if \code{year},
\code{month}, and \code{dekad} are not specified.}

\item{year}{\code{character(1)} Name of the year column in \code{data}. Only needed if \code{type
  = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::year(data[[date_time]])}.}

\item{month}{\code{character(1)} Name of the month column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using
\code{lubridate::month(data[[date_time]])}.}

\item{dekad}{\code{character(1)} Name of the dekad column in \code{data}. Only needed if
\code{type = "dekad"}. If \code{NULL} it will be created using \code{dekad}
function.}

\item{metadata}{\code{data.frame} data.frame of station metadata. Use this is the station
details are in a separate data.frame with one row per station. If
specified, \code{latitude}, \code{longitude} and \code{altitude} are
assumed to be in \code{metadata} and \code{station} must be in both
\code{data} and \code{metadata} to facilitate joining.}

\item{latitude}{\code{character(1)} Name of the latitude column in \code{metadata}.}

\item{longitude}{\code{character(1)} Name of the longitude column in \code{metadata}.}

\item{altitude}{\code{character(1)} Name of the altitude column in \code{metadata}.}
}
\value{
A data.frame formatted for use in CDT.
}
\description{
Prepare data in format for CDT
}
\examples{
# Create prepare summary dekad data for CDT export
summary_data <- daily_niger \%>\%
                   dplyr::mutate(dekad_date = dekad(daily_niger$date)) \%>\%
                   dplyr::group_by(station_name, year, dekad_date) \%>\%
                   dplyr::summarise(date = dplyr::first(date), sum = sum(tmax))

prepare_cdt_dekad(data = summary_data, date_time = "date", year = "year",
                  station = "station_name",
                  element = "sum", metadata = stations_niger, 
                  latitude = "lat", longitude = "long", altitude = "alt")
}
