% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_climdex.R
\name{prepare_climdex}
\alias{prepare_climdex}
\title{Prepare ClimDex data for export}
\usage{
prepare_climdex(
  data,
  prcp,
  tmax,
  tmin,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{prcp}{The name of the rainfall column in \code{data}.}

\item{tmax}{The name of the maximum temperature column in \code{data}.}

\item{tmin}{The name of the minimum temperature column in \code{data}.}

\item{date}{The name of the date column in \code{data}. This is only needed if \code{year}, \code{month}, and \code{day} are not specified.}

\item{year}{The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date]])}.}

\item{month}{The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date]])}.}

\item{day}{The name of the day column in \code{data}. If \code{NULL} it will be created using \code{lubridate::day(data[[date]])}.}
}
\value{
Invisibly returns the file path of the saved data
}
\description{
Prepare ClimDex data for export
}
\examples{
# Preparing the daily_niger data for export to ClimDex
prepare_climdex(data = daily_niger, date = "date", prcp = "rain", tmax = "tmax", tmin = "tmin")

}
