% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_climdex.R
\name{prepare_climdex}
\alias{prepare_climdex}
\title{Prepare data in the format for RClimDex}
\usage{
prepare_climdex(
  data,
  prcp,
  tmax,
  tmin,
  date = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  na = -99.9
)
}
\arguments{
\item{data}{\code{data.frame} data.frame of daily climatic data in tidy format i.e. one row per
day and one column per element.}

\item{prcp}{\code{character(1)} Name of the precipitation/rainfall column in \code{data}.}

\item{tmax}{\code{character(1)} Name of the maximum temperature column in \code{data}.}

\item{tmin}{\code{character(1)} Name of the minimum temperature column in \code{data}.}

\item{date}{\code{character(1)} Name of the date column in \code{data}. This is only needed if
\code{year}, \code{month}, and \code{day} are not specified.}

\item{year}{\code{character(1)} Name of the year column in \code{data}. If \code{NULL} it will be
created using \code{lubridate::year(data[[date]])}.}

\item{month}{\code{character(1)} Name of the month column in \code{data}. If \code{NULL} it will
be created using \code{lubridate::month(data[[date]])}.}

\item{day}{\code{character(1)} Name of the day of the month column in \code{data}. If \code{NULL}
it will be created using \code{lubridate::day(data[[date]])}.}

\item{na}{\code{integer(1)} The value to replace missing values with. The default (-99.9)
should be used to comply with the RClimDex format requirements.}
}
\value{
A data.frame formatted for use in RClimDex
}
\description{
\code{prepare_climdex} rearranges a data frame to a
format suitable for use in RClimDex. This data frame can be rearranged and
exported to a file or connection with \code{export_climdex}.
}
\examples{
# Preparing the daily_niger data for export to ClimDex
prepare_climdex(data = daily_niger, date = "date", prcp = "rain", tmax = "tmax", tmin = "tmin")

}
