% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim.R
\name{prepare_geoclim}
\alias{prepare_geoclim}
\title{Prepare dekad or pentad data in GeoCLIM format}
\usage{
prepare_geoclim(
  data,
  year,
  type_col,
  element,
  station_id,
  latitude,
  longitude,
  type = c("dekad", "pentad"),
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{type_col}{\code{character(1)} The name of the dekad or pentad column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{\code{character(1)} The name of the station column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{type}{\code{character(1)} Whether the data is in \code{dekad} or \code{pentad} format.}

\item{metadata}{\code{data.frame} The metadata data.frame to calculate from.}

\item{join_by}{\code{character} The variable(s) to merge the \code{data} and \code{metadata} data frames.}

\item{add_cols}{\code{character} Names of additional metadata columns that should be included in the output}
}
\value{
A data.frame formatted for use in GeoCLIM
}
\description{
Prepare dekad or pentad data in GeoCLIM format
}
\examples{
# Calculate dekadal summaries for the rainfall column
dekad_data <- daily_niger \%>\% dplyr::mutate(dekad = dekad(date))
summary_data <- dekad_data \%>\% dplyr::group_by(station_name, year, dekad) \%>\%
      dplyr::summarise(mean_rain = mean(rain, na.rm = TRUE))
prepare_geoclim(data = summary_data, year = "year",
                station_id = "station_name",
                type_col = "dekad",
                element = "mean_rain", metadata = stations_niger, 
                join_by = "station_name",
                latitude = "lat", longitude = "long")
}
