% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim.R
\name{prepare_geoclim}
\alias{prepare_geoclim}
\title{Prepare dekad or pentad data in GeoCLIM format}
\usage{
prepare_geoclim(
  data,
  year,
  type = c("dekad", "pentad"),
  type_col,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL
)
}
\arguments{
\item{data}{}

\item{year}{}

\item{type}{}

\item{type_col}{}

\item{element}{}

\item{metadata}{}

\item{join_by}{}

\item{station_id}{}

\item{latitude}{}

\item{longitude}{}

\item{add_cols}{}
}
\value{

}
\description{
Prepare dekad or pentad data in GeoCLIM format
}
