% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_dekad.R
\name{prepare_geoclim_dekad}
\alias{prepare_geoclim_dekad}
\title{Prepare dekad data in GeoCLIM format}
\usage{
prepare_geoclim_dekad(
  data,
  year,
  dekad,
  element,
  station_id,
  latitude,
  longitude,
  metadata = NULL,
  join_by = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{dekad}{\code{character(1)} The name of the dekad column in \code{data}.}

\item{element}{\code{character(1)} The name of the element column in \code{data} to apply the function to.}

\item{station_id}{\code{character(1)} The name of the station column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{latitude}{\code{character(1)} The name of the latitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{longitude}{\code{character(1)} The name of the longitude column in \code{metadata}, or \code{data} if \code{metadata = NULL}.}

\item{metadata}{\code{data.frame} The metadata data.frame to calculate from.}

\item{join_by}{\code{character} The variable(s) to merge the \code{data} and \code{metadata} data frames.}

\item{add_cols}{Names of additional metadata columns that should be included in the output}
}
\value{
A data.frame formatted for use in geoclim.
}
\description{
Prepare dekad data in GeoCLIM format
}
\examples{
# TODO
}
