% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_month.R
\name{prepare_geoclim_month}
\alias{prepare_geoclim_month}
\title{Prepare monthly data in GeoCLIM format}
\usage{
prepare_geoclim_month(
  data,
  year,
  month,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL
)
}
\arguments{
\item{data}{}

\item{year}{}

\item{month}{}

\item{element}{}

\item{metadata}{}

\item{join_by}{}

\item{station_id}{}

\item{latitude}{}

\item{longitude}{}

\item{add_cols}{}
}
\value{

}
\description{
Prepare monthly data in GeoCLIM format
}
