% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geoclim_pentad.R
\name{prepare_geoclim_pentad}
\alias{prepare_geoclim_pentad}
\title{Prepare pentad data in GeoCLIM format}
\usage{
prepare_geoclim_pentad(
  data,
  year,
  pentad,
  element,
  metadata = NULL,
  join_by = NULL,
  station_id,
  latitude,
  longitude,
  add_cols = NULL
)
}
\arguments{
\item{data}{}

\item{year}{}

\item{pentad}{}

\item{element}{}

\item{metadata}{}

\item{join_by}{}

\item{station_id}{}

\item{latitude}{}

\item{longitude}{}

\item{add_cols}{}
}
\value{

}
\description{
Prepare pentad data in GeoCLIM format
}
