% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plot.R
\name{timeseries_plot}
\alias{timeseries_plot}
\title{Title}
\usage{
timeseries_plot(
  data,
  time,
  elements,
  station = NULL,
  facets = c("none", "elements", "stations", "both"),
  add_points = FALSE,
  add_line_of_best_fit = FALSE,
  se = TRUE,
  add_path = FALSE,
  add_step = FALSE,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{data}{}

\item{time}{}

\item{elements}{}

\item{station}{}

\item{facets}{}

\item{add_points}{}

\item{add_line_of_best_fit}{}

\item{se}{}

\item{add_path}{}

\item{add_step}{}

\item{na.rm}{}

\item{show.legend}{}
}
\value{

}
\description{
Title
}
