% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plot.R
\name{timeseries_plot}
\alias{timeseries_plot}
\title{Plot a timeseries graph using \code{ggplot2}}
\usage{
timeseries_plot(
  data,
  date_time,
  elements,
  station = NULL,
  facets = c("stations", "elements", "both", "none"),
  type = c("line", "bar"),
  add_points = FALSE,
  add_line_of_best_fit = FALSE,
  se = TRUE,
  add_path = FALSE,
  add_step = FALSE,
  na_rm = FALSE,
  show_legend = NA
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{The name of the date column in \code{data}.}

\item{elements}{The name of the column in \code{data} to apply the function to.}

\item{station}{The name of the station column in \code{data}, if the data are for multiple station.
Timeseries plots are calculated separately for each station.}

\item{facets}{How to split the time series plots. Default \code{"station"}. Can be one of \code{"elements"}, \code{"both"}, or \code{"none"}.}

\item{type}{The type of plot, either "line" or line graphs or "bar" for bar graphs.}

\item{add_points}{logical. If \code{TRUE}, points are added to the plot using  \code{"ggplot2::geom_point()"}.}

\item{add_line_of_best_fit}{logical. If \code{TRUE}, points are added to the plot using  \code{"ggplot2::geom_smooth(method = "lm")"}.}

\item{se}{logical. If \code{TRUE}, the standard error is are added to the line of best fit. Only works if \code{add_line_of_best_fit = TRUE}.}

\item{add_path}{logical. If \code{TRUE}, paths are added to the plot using  \code{"ggplot2::geom_path()"}.}

\item{add_step}{logical. If \code{TRUE}, steps are added to the plot using  \code{"ggplot2::geom_step()"}.}

\item{na_rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show_legend}{logical. Should this layer be included in the legends? \code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes.}
}
\value{
A plot of type ggplot to the default plot device
}
\description{
Creates a timeseries graph for each element and station given. Takes a data frame as an input and the relevant columns to create the plot.
Creates a graph using \code{ggplot2} and returns a timeseries plot.
}
\examples{
# Create a time series plot with two elements and facet by station.
data(daily_niger)
timeseries_plot(data = daily_niger, date_time = "date", elements = c("tmax", "tmin"), 
                station = "station_name", facets = "stations")

# Can make additional changes to the plot since the returned object is a \code{ggplot2} object
# for example, to add colour-blind friendly colours instead of the default colours
require(ggplot2)
t1 <- timeseries_plot(data = daily_niger, date_time = "date", elements = c("tmax", "tmin"), 
                      station = "station_name", facets = "stations")
t1 + ggplot2::scale_colour_discrete(type = c("#E69F00", "#56B4E9"))
}
