% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Windrose from the clifro package}
\usage{
windrose(
  data,
  speed,
  direction,
  facet_by = NULL,
  n_directions = 12,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  ggtheme = c("grey", "gray", "bw", "linedraw", "light", "minimal", "classic"),
  legend_title = "Wind Speed",
  calm_wind = 0,
  variable_wind = 990,
  n_col = NULL
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{speed}{\code{numeric} A vector containing wind speeds.}

\item{direction}{\code{numeric} A vector containing wind direction.}

\item{facet_by}{\code{character(1)} Facets used to plot the various windroses.}

\item{n_directions}{\code{integer(1)} The number of direction bins to plot (petals on the rose).
The number of directions defaults to 12.}

\item{n_speeds}{\code{numeric(1)} The number of equally spaced wind speed bins to plot.
This is used if speed_cuts is NA (default 5).}

\item{speed_cuts}{\code{numeric} A vector containing the cut points for the wind speed intervals, or NA (default)}

\item{col_pal}{\code{character} String indicating the name of the \code{RColorBrewer} colour palette to be used for plotting.}

\item{ggtheme}{\code{character(1)} String (partially) matching the \code{\link[ggplot2]{ggtheme}} to be used for plotting.}

\item{legend_title}{\code{character(1)} Legend title.}

\item{calm_wind}{\code{numeric(1)} The upper limit for wind speed that is considered calm (default 0).}

\item{variable_wind}{\code{numeric(1)} Variable winds (if applicable).}

\item{n_col}{\code{integer(1)} The number of columns to plot (default 1).}
}
\value{

}
\description{
A wrapper for plotting a windrose of wind speed and direction using \code{ggplot2}
}
\examples{
# TODO
}
