% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{Windrose from the clifro package}
\usage{
windrose(
  data,
  speed,
  direction,
  facet = NULL,
  n_directions = 12,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  ggtheme = c("grey", "gray", "bw", "linedraw", "light", "minimal", "classic"),
  legend_title = "Wind Speed",
  calm_wind = 0,
  variable_wind = 990,
  n_col = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{speed}{TODO}

\item{direction}{TODO}

\item{facet}{TODO}

\item{n_directions}{TODO}

\item{n_speeds}{TODO}

\item{speed_cuts}{TODO}

\item{col_pal}{TODO}

\item{ggtheme}{TODO}

\item{legend_title}{TODO}

\item{calm_wind}{TODO}

\item{variable_wind}{TODO}

\item{n_col}{TODO}
}
\value{

}
\description{
A wrapper for plotting a windrose of wind speed and direction using ggplot2
}
\examples{
# TODO
}
