% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwr_export.R
\name{wwr_export}
\alias{wwr_export}
\title{Title}
\usage{
wwr_export(
  data,
  year,
  month,
  mean_station_pressure,
  mean_sea_level_pressure,
  mean_temp,
  total_precip,
  mean_max_temp,
  mean_min_temp,
  mean_rel_hum,
  link,
  link_by,
  station_data,
  wmo_number,
  latitude,
  longitude,
  country_name,
  station_name,
  height_station,
  height_barometer,
  wigos_identifier,
  folder
)
}
\arguments{
\item{data}{The data.frame to calculate from}

\item{year}{}

\item{month}{}

\item{mean_station_pressure}{}

\item{mean_sea_level_pressure}{}

\item{mean_temp}{}

\item{total_precip}{}

\item{mean_max_temp}{}

\item{mean_min_temp}{}

\item{mean_rel_hum}{}

\item{link}{}

\item{link_by}{}

\item{station_data}{}

\item{wmo_number}{}

\item{latitude}{}

\item{longitude}{}

\item{country_name}{}

\item{station_name}{}

\item{height_station}{}

\item{height_barometer}{}

\item{wigos_identifier}{}

\item{folder}{}
}
\value{

}
\description{
Reshape data into formats required by WMO for submission of climatic data
this gives Yearly data records with monthly and annual data for a particular year.
}
